Create harvestor.ps1
Initial release. From the file containing URLs (sitemap.xml), each page is loaded and based on the pattern in InnerText, the content of the tag is gathered.
It just suffer of the duplicate items, if you scrape e.g. huge site, where at each footer is the same e-mail address placed at. However, this might be easily fixed e.g. in Excel, where the resulting file is opened at.